requirements.txt
flask
dnspython
python-whois
requests
python-Levenshtein
reportlab